Support JSON in the broker status. Enable inet sockets everywhere.
The broker status socket now expects to receive JSON which is parsed
using the shiny new pstore-json library. Enable internet sockets on
all platforms rather than just Windows (using a new --inet switch on
both server and client); this should enable some system
testing of the infrastructure.

+++ lib/support/file_posix.cpp
@@ -386,10 +386,7 @@ namespace pstore {
         namespace posix {
 
             void deleter::platform_unlink (std::string const & path) {
-                if (::unlink (path.c_str ()) == -1) {
-                    int const err = errno;
-                    raise_file_error (err, "unlink failed", path);
-                }
+                pstore::file::unlink (path);
             }
 
         } // namespace posix
@@ -408,10 +405,12 @@ namespace pstore {
             }
         }
 
-        void unlink (std::string const & path) {
+        void unlink (std::string const & path, bool allow_noent) {
             if (::unlink (path.c_str ()) == -1) {
                 int const err = errno;
-                raise_file_error (err, "unlink failed", path);
+                if (!allow_noent || err != ENOENT) {
+                    raise_file_error (err, "unlink failed", path);
+                }
             }
         }
     } // namespace file
